pyproject.toml
[project]
name = "binance-trading-bot"
version = "0.1.0"
description = "Automated cryptocurrency trading bot for Binance with API interface"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-binance>=1.0.29",
    "python-dotenv>=1.1.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "websocket-client>=1.6.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "rich>=14.0.0",
]
